Do not build HTML versions of the compendium by default.
Due to problems with hyperlatex and emacs23, Debian does not
distribute hyperlatex with Squeeze.  We need hyperlatex only for the
HTML versions of the manual, which are anyway not included in the
installer.  Thus we take a simple way out and include the hypertex.sty
file in gpg4win for building the PDF version.

* doc/manual/hyperlatex.sty: New.  File taken from Debian's hyperlatex
version 2.9a.
* doc/manual/Makefile.am (EXTRA_DIST): Distribute hyperlatex.
(all-local): Remove the html target.