Disable AES-NI support in Libgcrypt.
--
There is a problem wit VirtualBox when using AES-NI.  Disabling
AES-NI for gpg-agent seems to be sufficient but we better play safe
and disable it completely.

Note that there is no problem with the Libgcrypt tests programs; even
the backported alignment test for 1.6 works fine.  Needs to be
investigated further.