Update GTK+ related packages.
* packages/packages.current: Update glib, gdk_pixbuf, gtk+, libffi and
check all other packages for necessary updates.
* patches/glib-2.34.3/01-socket.patch: Copy from 2.24.0.
* patches/glib-2.34.3/02-no-kill.patch: New.
* Makefile.am (EXTRA_DIST): Add new patches.
* src/Makefile.am (gpg4win_pkg_glib_configure): Add
option --disable-modular-tests.
* src/inst-libffi.nsi: Update SO-name from 5 to 6.
* src/uninst-libffi.nsi: Ditto.  Remove unneeded top dir rmdir.
--

The update of glib from 2.32 to 2.34 is due to a changed requirement
in gdk_pixbuf 2.26.5 (which has been updated from 2.26.0).

Note that the GPA crashes in the latest beta are due to the missing
01-socket.patch.  I simply forgot to copy it to the newer gtk version.
We should tell the GTK folks to include our patches.