Fix a hang in gpgme.
* patches/gpgme-1.4.1/01-improve-trace.patch: New.
* patches/gpgme-1.4.1/02-close-socket.patch:
* Makefile.am (EXTRA_DIST): Add patches.
--

When using a socket (e.g. polling scdaemon) the closing of the socket
often blocked waiting on termination of the reader thread.  At that
point the reader table lock is also held and thus no other reader
thread could be started.  See the comments in the code for out
solution.

Note that there is currently another hang in Kleopatra, the first time
it is started. I was able to workaround that by disabling the startup
selftests.