Temporary fix for mingw off_t mess.
* patches/gpgol-1.1.3/02-off_t.postcfg-build: New.
* Makefile.am (EXTRA_DIST): Add patch
--

The problem with w64-mingw32's off_t hack is that it is incompatible
to Windows.  It works nicely with most software but the DLL's using
off_t in the public API (e.g. libgpgme) can't be used by MSVC build
programs.  That is a regression.

The tentative plan to fix this in gpgme is to use int32_t instead of
off_t on Windows.