MSI: Add build system integration for Wixtools
* src/Makefile.am (msi): Change target to work with
wixtools.

--
This builds an MSI package. Currently by building
a wixlib for Gpg4win, too and linking together
the wixlibs on Windows.

Windows appears to be required because even though
we were able to create a working MSI for GnuPG using
wine and wixtools on Linux, for Gpg4win it failed
with strange errors that were not easy to figure
out.