MSI: Add comment why selfreg is required
* src/make-msi.pl: Add comment why selfreg is required.

--

Selfreg causes a DLL to be registered with the Windows regsvr
basically DLLInstallServer is called in the dll. This is not
considered to be good style in an MSI package but the alternative
for us would be to create a 64 bit and a 32 bit package, which
we prefer to avoid.