MSI: Print warning about x86 MSI installers
* src/gpg4win.mk.in: Print warning for x86 installer.

--
I leave this a fixme for now. I don't understand why components
marked with Win64=Yes are selected in the 32 bit installer. This
worked when the package element defined the architecture. It is
more important that the 64 bit installer works so for now
let us leave this a fixme and hope that 32 bit windows goes away
before we have to fix this ;-).