Remove gpgwrap installation
* src/inst-gnupg-w32.nsi, src/uninst-gnupg-w32.nsi:
Remove gpgwrap installation.

--
We currently do not use them anymore and they were more
of an experiment to avoid windows when running from
login scripts.
We disable them for now as at least two anti "virus" scanners
detect them as malware, which is kind of understandable in
the way that they just wrap a different process.