Add/update some packages to download
* packages/packages.current: Add SQLite. Add info about Werner's new
dist signing key. Add npth, libgcrypt, libksba, gnupg. Update gpgme.
Add qtx11extras, qtwayland, kdbusaddons. Add latest released versions
of libkleo and kleopatra. Change server URLs to use https. Disable
usage of snapshots of gpgme, libkleo, kleopatra.
--

SQLite and the GnuPG source packages are needed because (currently) we
do not use a binary built of gpg for the AppImage.
qtx11extras, qtwayland, kdbusaddons are needed for building pinentry,
some KDE frameworks, and kleopatra for the AppImage.
The snapshots of libkleo and kleopatra were not compatible (libkleo is
too old) and the snapshot of kleopatra doesn't compile with Qt 5.15.0.
Therefore, I switched to the release versions for the time being.

GnuPG-bug-id: 5592