Copy Qt translation files to the folder expected by Qt/Kleopatra
* docker/appimage/build-appimage.sh: Remove existing Qt translations
before (re)running linuxdeploy.
* src/appimage/Makefile.am (appimage_pkg_qttranslations_post_install):
Install translations in "translations" subfolder.
--

On Linux, Qt expects the translations in the "translations" subfolder.
The linuxdeploy Qt plugin copies the necessary translation files to the
AppDir. It also tries to create bogus symlinks. Removing existing
Qt translations from AppDir before (re)running linuxdeploy is a
workaround for this ideosyncrazy of linuxdeploy. It will still try to
create those symlinks, but this fails with non-fatal errors which we
have to ignore for now.

GnuPG-bug-id: 5592