Add a script for creating a snapshot of a package
* packages/gen-tarball.sh: New.
--

This script packages a snapshot of the current HEAD of a git repository
as a tarball and generates a suitable text snippet for packages.current.
Expects the name of a supported package (currently only kleopatra is
supported) or the local path to a git repository. In the latter case,
the basename of the path is used as name of the tarball.