appimage: Allow building a VSD version and rename AppImage name
* docker/appimage/build-appimage.sh: Detect VSD build and set sysconf
dir to /etc.
* src/appimage/AppRun: Prettify
--

We detect a VSD build by the presence of a gnupg-vsd sub-directory in
the same way a gpg4win dies this.

The AppImage name is also changed to either gnupg-desktop-2.n.m-* or,
for a VSD build, to gnupg-vs-desktop-3.n.m-*.  The version number for
the latter is taken from Kleopatra because that reflects the actual
version in this case.

The sysconfdir is now fixed to /etc because a sysconfdir inside the
AppImage does not make much sense.  For a VSD build a dedicated name
is used so that this does not conflict with the system's default use
of gpg.

Signed-off-by: Werner Koch <wk@gnupg.org>