Disable unpack and patch for symlinked dirs
* src/gpg4win.mk.in (stamp-unpack, stamp-patch): Disable actions
when on symlinked dirs.
(stamp-ex-configure): Avoid error if path already exists.

--
This is a quality of life patch for me because when
developing I usually symlink my development repos into
the build tree of Gpg4win. Not having to redo this after
each removal / update of the relevant stamps is helpful.