MSI: Split into multiple cabs and reuse them
* src/make-msi.pl (MediaTemplate): Split into cabs of
10MB uncompressed data.
* src/gpg4win.mk.in: Add -cc and -reuscab options.

--
Only using -cc (Cache Path) and -reusecab did not help as
by default we only created a single cabinet file. As this
changes each time it would not be reused.

Automatic splitting is a compromise between working
with Media attributes which would have needeed to be
added to each Component etc. Media would have the advantage
that we could only mark the dynamic parts as a second media.

For me the first that created the Cabinets: 45.309 total
Second run with cached Cabinets: 24.200 total