Remove mkportable.exe
* mkportable-full.h, mkportable-vanilla.h, mkportable.c: Removed.
* inst-gpg4win.nsi, uninst-gpg4win.nsi: Do not install / uninstall
mkportable.exe
* Makefile.am: Update accordingly.

--
mkportable.exe has not really worked as advertised since the
KDE 5 Update, (Gpg4win-3.0.0) since Kleopatra would
use configuration files from %LOCALAPPDATA% an %APPDATA%.

Additionally in many releases the file lists were not up
to date causing errors on copy.

While we still think it is usually not a good idea
to have a portable installation, it is still documented
how to do this for GnuPG in the GnuPG Documentation under
--homedir.

As for Gpg4win you can just copy the files anywhere it
can then also find GnuPG e.g. in the same directory.

GnuPG-Bug-Id: T6246