build: New script to ease packaging of gnupg
--

Often GnuPG is build on the same machine as GnuPG-Desktop.  This
scripts makes it easy to add a new GnuPG version to the packages file.
Simply remove old gnupg versions:

  cd packages
  rm gnupg-*

run

  ./gen-gnupg.sh ~/b/gnupg/dist/gnupg-w32....tar.xz

so that the scripts copies all needed files from the gnupg build
directory to the local director.  Finally edit packages.N and delete
the old version info.