Update Kleopatra and deps to latest master
* Makefile.am (EXTRA_DIST): Update KCMUtils patch.
* packages/packages.common (gpgme, mimetreeparser, libkleo)
(kleopatra): Update to latest master.
* patches-appimage/kleopatra/0001-Make-KCMUtils-optional.patch,
patches/kleopatra/0001-Make-KCMUtils-optional.patch: Rewrite
to be more clean instead of the old "revert" patch.
* patches/kleopatra/set-windows-registry.patch: Refresh.

--
The KCMUtils patch did no longer apply with the clang format
changes so I modified it in a way that might even be feasible
for upstream kleopatra. Since it apparently also only has an
effect for Qt6.