Prepare registry rules for GpgEX replacement
* src/inst-kleopatra.nsi: Add a commented out entry for
all files to sign/encrypt.

--
This is the result of some experiments to replace GpgEX by some
registry entries. This works but is commented out to avoid a
mess while GpgEX is still installed and this is not yet complete.
I just wanted to commit it so it is not lost.

GnuPG-Bug-Id: T6331