Fix S/MIME file extensions
* src/inst-kleopatra.nsi, src/uninst-kleopatra.nsi
  (gpg4win.AssocFile.kleopatra.X509): Register for all cert
  types. Remove AllowSilendDefaultTakeOver.
* src/make-msi.pl: Remove AllowSilentDefaultTakeOver.

--
The allowSilentDefaultTakeOver was the bad point. I thought
it prevented Kleopatra from taking over .p12 files directly while
in fact it prevented the nice user friendly: "Hey there is
a new Application that can handle this type of file popup,
do you want to use this one." Popup...

This works as intended now. Kleopatra is shown as the first alternative
for all the files where Windows already has builtin support for.