Install breeze icons as rcc file
* Makefile.am (EXTRA_DIST): Add new patch.
* patches/breeze-icons/add-the-nose.patch: New. Replace the
Breeze Kleo icon with our Gpg4win one.
* src/included-icons.txt: Add kleopatra
* src/inst-breeze-icons.txt, src/uninst-breeze-icons.nsi: Only
install the icontheme.rcc files.
* patches/patches/qtbase/0002-Gpg4win-theme-names-and-relpaths.patch:
Remove custom theme / icon code.

--
This adds the Kleopatra icon to breeze because we then have no
need anymore to look up icons in the highcolor theme. Ok that
looses the logo for the Gpg4win compendium but I can live
with that. The original breeze-icon for Kleopatra was not
adopted by Gpg4win because the Kleopatra icon was a discussion
of epic proportions and we did not want to open that can of
worms again.

This should finish the icon side of:
GnuPG-Bug-Id: T4066