doc: Remove README files
* doc/README*: Removed.
* src/Makefile.am (README_files): Removed.
* src/gpg4win.mk.in: Remove readme template.
* src/inst-gpg4win.nsi, src/uninst-gpg4win.nsi: Remove mention
of README files.
* src/installer.nsi: Remove link to open the README.

--
The README files were generally outdated and did not contain
any useful information that is not better available from
other sources. At least for a GUI software a text based README
is no longer appropiate in my opinion. If there is important
information that users need to know the Software should present
that itself.

GnuPG-Bug-Id: T6779