Add kcoreaddons patches for QSettings based about data
* Makefile.am (EXTRA_DIST): Add patches.
* patches/patches/kcoreaddons/
0001-Add-KAboutData-updateFromSettings.patch: New. Ready for review.
* patches/kcoreaddons/
0002-Draft-Add-KVerify-class-to-verify-files.patch: New. But needs
some more documentation and maybe extension in usefulness.
* patches-appimage/kcoreaddons: New symlink to windows patches.

--
For now it is easier to patch this then to put up a branch
and keep it updated. The AboutData patch is IMO ready for
a review and I will put it up as MR. The KVerify patch needs
a bit more work.

GnuPG-Bug-Id: T6594