MSI: Kill running processes
* src/gpg4win.mk.in: Link in WixUtilExtension.
* src/make-msi.pl (product): Hardcode the kill sequence.

--
WixUtil is required for WixQuietExec.

This sets up custom actions and the according sequence to
use WixQuietExec to kill our running background processes.
It is done twice because it is first run before the elevation
and then a second time in the main install sequence after
the UAC elevation. That was the reason while failures to kill
always were reported twice.

While this fixes the upgrade with running backgorund processes
it will also fix that the Windows Explorer is properly restarted
after Upgrading GpgOL. Since I realized that if the RestartManager
failed to kill some processes it will just bail out and not
restart anything after an upgrade. Now both Outlook and the
Windows Explorer are restarted if they are running during an
Upgrade.

I have kept the sequence hardcoded instead of adding some perl
code that would just take a list of processes to better be able
to comment this. It should be obvious how to add new tasks to
be killed.

GnuPG-Bug-Id: T6567