Update Okular about data patch. To read from ini.
--
This replaces the patch that hardcoded the values for the GnuPG
Edition of Okular in the aboutdata and instead lets us choose a file
depending on the language of the installer. It currently
does not a verify check since I am a bit worried about what is then
updated at which point. In Kleopatra the verification seems more
important to me.

GnuPG-Bug-Id: T6594