Install kcoreaddons license texts
* src/inst-kcoreaddons.nsi, src/uninst-kcoreaddons.nsi: Install/
Uninstall license texts.

--
This automatically makes the license link in the about dialog
of Kleopatra and Okular work and covers the "You should have
received the license along with this program"

KCoreAddons is a good place because the GnuPG installer installs
its own licenses and everything else is then just dependencies for
KDE based Software.

GnuPG-Bug-Id: T6834