Update copyright list.
--

For a couple of packages (in particular KDE stuff and poppler) the
list of copyright holders are missing or are incomplete.  Should be
fixed. For KDE stuff something like

  grep -Ri SPDX-FileCopy * | sed 's/^.*SPDX-.*: //' \
     | sort | uniq | sed 's/<.*>//' | sort | uniq

might be useful to a get a list and the manually merge entries.
Debian should also have toold to figure out copyright holders.