Include all code to sign MSI files.
* src/Makefile.am: Remove optional include of sign.mk.
(AUTHENTICODE_FILES): New.
(READ_AUTOGEN_template): New.  Also read a couple of vars.
(msi_signkey): New.
* src/gpg4win.mk.in (AUTHENTICODE_sign): Replace timestamp URL by a
variable.
--

This code formerly lived in an external file but we decided that it is
safer to have it in gpg4win proper.  We now also use
~/.gnupg-autogen.sh to get custom values like URLs. Keys and tool
locations.  A description of this file can be found in
gnupg/build-aux/speedo.mk.