Use a separate script to authenticode sign files.
* build-aux/authenticode-sign.sh: New.
* src/Makefile.am (stamp-additional-signedfiles): Call this tool.
* src/gpg4win.mk.in: Remove most of calls of READ_AUTOGEN_template.
(AUTHENTICODE_sign): Remove this macro.
(signed_installers): Call the new tool
--

The problem with the old use was that it was used in one place in an
eval and at another place directly.  Thus the quoting required to be
different.  A separate tools sounds better and can be extended for
other purposes.  Note the tool's option --template.

Only the card based operation has been tested; the old signing on a
Windows host might be broken.