Add zstd dependency for Qt
--
This enables the zstd feature for Qt6 and KArchive.
There is an implicit dependeny on that feature in
Frameworks.

Additionally it adds the variable:
pkg_conf_subdir="$(call GETVAR,gpg4win_pkg_$(1)_conf_subdir)"

to declare a non standard subdir to configure a project.
Currently only implemented for KDEPKGs (cmake)