Use qt.conf again to change QtStandardPaths
* patches/qtbase/0001-Gpg4win-qstandardpaths-patch.patch: Refreshed.
* src/inst-qtbase.nsi: Install qt.conf file instead of writing it.
* src/qt.conf: New.

--
When the environment variable GNUPGHOME is set this returns
the according values using QStandardPaths. The basic patch for
this has been around since Qt5 (or even 4) to avoid conflicts
with other distributions of KDE Software on Windows.

This was not applied for GnuPG VS-Desktop since it used
the FileWrite mechanism of NSIS instead if installing the
file, which is not converted into WXS.

GnuPG-Bug-Id: T6799