docker: Rewrite docker build scripts
* docker/docker-build.sh: New.

--
The new docker-build.sh replaces the old scripts and
provides more flexibility to work with Gpg4win. Basically
creating a way for out of tree builds.

Default call to docker/docker-build.sh clones the
current repository as the source.
To rebuild a release tarball use the --inplace command.

Build Gpg4win in a docker containter.

Options:
        --appimage      Build the AppImage
        --gpg-2.2       Use GnuPG 2.2 instead of the default
        --dirty         Include uncommited changes
        --shell         Start a shell instead of starting a buildscript
        --root-shell    Start a root shell
        --clean-pkgs    Do not use already downloaded packages
        --inplace       Build in the current directoy
        --buildroot     Directory where the build should take place

This builds either the Appimage or Gpg4win for Windows. To
build a source tarball the option inplace can be used. By
default it builds in a temporary directory use the
option --inplace or --build-dir to change that behavior.