Update gettext and iconv
* packages/packages.common (gettext, iconv): Update.

--
This was done in an attempt to fix the
"checking for working iconv" test of GnuPG. But
in the end the reason for that error was that
glibc-locale-base was not installed on the
build system.

Updating these packages as general maintenance
makes sense now that we have a major library
upgrade anyway.