Depend gpg4win version on gpg version again
* configure.ac: Check for buildtype and for
GnuPG Version to see which version should be used.

--
If we are using GnuPG 2.2.x or building GnuPG VSD then
use a different version number. This allows configurations
where the buildtype is still VSD but the version used
for GnuPG might differ e.g 2.5 or 2.6.

But for the default build since we offer two flavors
depending on the downloaded versions we need to
distinguish those, otherwise the situation where
the same Gpg4win installer / version is built with
different GnuPG versions. Which can lead to confusion
if it is done accidantlly.