MSI: Replace obsolete PRODUCT macros
* src/gpg4win.mk.in (VSD_PRODUCT, VSD_PRODUCT_PRETTY): Replace by
MSI_FNAME_PREFIX and MSI_PRODUCT respectively.

--
The complex logic with the $$($(1)_PRODUCT) macro was
required to support different products in the same
build run. This was never really used and we even
use different builds for GPD and VSD nowadays.

This simple replacement of $$($(1)_PRODUCT) by
$(MSI_FNAME_PREFIX) and $$($(1)_PRODUCT_PRETTY) by
$(MSI_PRODUCT_NAME) also fixes the issue that the
wixlib target were redefined for several times
causing a bit of error SPAM.