Add qt patch to make default style configurable
* Makefile.am (EXTRA_DIST): Add new patch.
* patches/qtbase/configurable-windowstheme.patch: New.
* src/qt.conf: Configure breeze style.

--
This is based on an old Qt5 patch we had to add breeze
as the default style but now it is also read from the
qt.conf so that it is easier to make experiments or
change it for some.

Okular already set the style in the Application
but this would mean changing the style manually
in all our qt applications from pinentry to
keyresolver. So an installation specific configuration
seems more sensible.

GnuPG-Bug-Id: T6932