Add local translations for kleopatra
* packages/gen-tarball.sh: Use a variable local_l10n additionally to
custom-l10n.
* packages/kleopatra-24.05-de-full-translation.po: New.

--
The full translation is a result of extrating the messages
in the current branch and then msgmerge'ing the translations from
the 24.05 repository and the translations from summit and then
manually fixing up missing strings.
This should hopefully only be neccessary because of the
distance between the gpg4win branch and the current kde translations.

New strings which are also in kleopatra master will still also
be added. So hopefully this should not need to be updated
a second time. See the comment in gen-tarball.sh for additional
details.