Add missing icons for KF5 build
* src/included-icons.txt: Add new icons

--
To find the icons I have just executed the script src/findicons.sh
this needs to be run in a build environment where all source pacakges
are extracted under src/playground/install.

The script is horribly inefficent but reliable enough. I jsust start
it and forget about it and in the end of the day it has added
the new icons to included-icons.txt.

Please note that with KF6 we do not yet filter icons with the
tool sune created. In KF6 all icons are combined in a DLL and then
loaded in every process which uses them. Filtering the icons
for KF6 would drastically reduce the size of the libarary and would
improve the load times. That is why I link the corresponding ticket.

To understand the script just look into it.
To use it just run it without arguments after a successful build
of Gpg4win. Probably overnight ^^.

GnuPG-Bug-Id: T4066