Update gen-tarball to auto update more langs
* packages/gen-tarball.sh: Loop over languages when updating with
custon_l10n
(ingore-msgcat-error): Remove option. Always just warn.

--
This extends the gen-tarball script to also update
locatization from KDE svn for other languages then
german. As a test I have tried with french to check
that the all certificates group was also translated.

This also fixes the order in which msgcat merges
in the local translation to correctly extend the
existing translations.

This remove the ignore-msgcat-error option and just
always prints a warning instead.