Improve merging of translations
* packages/gen-tarball.sh: Remove obsolete messages before merging
po files. Re-add --use-first to msgcat.
--

This fixes the problem that msgcat --use-first seemingly dropped the
context-less messages from the local translations when merging them.
It turned out that the context-less messages were flagged as obsolete
in the new translations and that msgcat --use-first kept those obsolete
context-less messages which effectively removed/hid them. Removing the
obsolete messages from the new translations (they are anyway dropped
when the po files are compiled) ensures that the not-obsolete messages
from the merged project translations and local translations are not
shadowed by the obsolete messages. And at the same time --use-first
takes care of removing outdated (local) translations.

GnuPG-bug-id: 7388