appimage: Use custom Kleopatra icons for VSD and GPD
* src/appimage/build-appimage.sh: For vsd or gpd replace Breeze icons
of kleopatra with custom icons. Otherwise restore the Breeze icons.
* src/icons/kleopatra-gpd.svg, src/icons/kleopatra-vsd.svg: New.
--

kleopatra-vsd.svg is the red icon from the kleopatra sources that we
use for our Windows builds. kleopatra-gpd.svg is a "GnuPG blue" variant
of kleopatra-vsd.svg. In the default build we need to restore the Breeze
icons from INSTDIR because they may have been replaced by a previous
vsd/gpd build.

GnuPG-bug-id: 7509