build: Create a file with changed sources.
* po/Makefile.am: Write a list of files which may need to be
committed.
* Makefile.am (clean-local): Clean that file.
--

Because the sources are read-only we can't updated them and it is
anyway better not to do that w/o manual interaction.  The new file can
be used to keep track of files to be committed.  We should also
consider to add the make-msi.guid file to this list.