Rework dependencies between cross-build packages and native packages
--

Collect the packages to cross-build and packages to build natively in
a single list. Native packages are prefixed with "native-" in this list.
Dependencies of cross-build packages to native packages need to use this
prefix. For example, all packages that need the natively built
translation tools from qttools now depend on native-qttools.
qttranslations also needs to depend on the cross-build qttools (for now)
because qttranslations's cmake files look for Linguist.

Native packages added with GPG4WIN_NATIVEPKG don't need to use the
prefix because it is added implicitly. (Native packages always depend on
other native packages.)

For NSIS we need to filter out the native packages because they are only
needed for the build and don't need to be packaged.