Ensure that download.sh --update doesn't remove breeze-icons
* packages/download.sh: Add special case for breeze-icons package.
* packages/packages.common: List breeze-icons *after* breeze.
--

If the breeze package needs to be downloaded then it removes breeze*xz
(which includes the breeze-icons package). Listing the breeze-icons
package after the breeze package makes sure that the breeze-icons
package is (re)downloaded in case it was removed during the update of
the breeze package. And special-casing breeze-icons (similar to
gnupg-w32 and gnupg-msi) makes sure that the update of the breeze-icons
package doesn't remove the breeze package.