gpgoljs: changed handling of js/css files and added translations
--

currently, the web client of GpgOL/Web (aka gpgoljs) is a static
javascript and css file with a hash in their file names. because
the MSI builder has issues with wildcards, a mechanism was added
that uses an included file from the gpgoljs tarball containing
the actual file names. since this will only work with newer gpgoljs
tarballs that actually carry the include file, packages.list was
also updated.