msi: Use "Everyone" for sysconfdir's gcrypt permissions.
* src/make-msi.pl (dump_single_custom):  Use Everyone.
--

BUILTIN/Users is translated on a German installation and can thus not
be used.  Everyone is not translated.

The actual goal is that users are able to read files from that
directory but not create the directory or files below it.  Note that
if a user can't read the files it would not a real problem and
definitely better than being able to create or update a file.

Fixes-commit: 0496f61cf429528688798544cd7ba4db37ed7b67