qtbase: Support custom app-specific standard paths
--

Additionally to environment variables the placeholder ::APPNAME:: is
replaced with the application name in the customized standard locations.
Moreover, the names of enum values are now resolved using QMetaEnum to
fix a wrong mapping of some enum values to their enum name and to avoid
such problems in the future.

GnuPG-bug-id: 7717